refactor: lift operation/parameter naming to parser config from inside parser
