+++ packages/openapi-ts/src/generate/__tests__/types.spec.ts
@@ -60,7 +60,6 @@ describe('generateTypes', () => {
           type: 'User',
         },
       ],
-      operationIds: new Map(),
       server: 'http://localhost:8080',
       services: [],
       types: {},
