+++ packages/openapi-ts/src/openApi/__tests__/index.spec.ts
@@ -1,19 +1,16 @@
 import { afterEach, describe, expect, it, vi } from 'vitest';
 
-import type { Config } from '../../types/config';
-import { parse } from '..';
+import { type OpenApi, parse } from '..';
+import type { Config } from '../common/interfaces/config';
 import * as parseV2 from '../v2';
 import * as parseV3 from '../v3';
 
-vi.mock('../../utils/config', () => {
-  const config: Partial<Config> = {
-    services: {},
-    types: {},
-  };
-  return {
-    getConfig: () => config,
-  };
-});
+const config: Config = {
+  nameFn: {
+    operation: () => 'operation',
+    operationParameter: () => 'operationParameter',
+  },
+};
 
 describe('parse', () => {
   afterEach(() => {
@@ -23,7 +20,7 @@ describe('parse', () => {
   it('uses v2 parser', () => {
     const spy = vi.spyOn(parseV2, 'parse');
 
-    const spec: Parameters<typeof parse>[0] = {
+    const spec: OpenApi = {
       info: {
         title: 'dummy',
         version: '1.0',
@@ -31,10 +28,10 @@ describe('parse', () => {
       paths: {},
       swagger: '2',
     };
-    parse(spec);
+    parse({ config, openApi: spec });
     expect(spy).toHaveBeenCalledWith(spec);
 
-    const spec2: Parameters<typeof parse>[0] = {
+    const spec2: OpenApi = {
       info: {
         title: 'dummy',
         version: '1.0',
@@ -42,14 +39,14 @@ describe('parse', () => {
       paths: {},
       swagger: '2.0',
     };
-    parse(spec2);
+    parse({ config, openApi: spec2 });
     expect(spy).toHaveBeenCalledWith(spec2);
   });
 
   it('uses v3 parser', () => {
     const spy = vi.spyOn(parseV3, 'parse');
 
-    const spec: Parameters<typeof parse>[0] = {
+    const spec: OpenApi = {
       info: {
         title: 'dummy',
         version: '1.0',
@@ -57,10 +54,10 @@ describe('parse', () => {
       openapi: '3',
       paths: {},
     };
-    parse(spec);
+    parse({ config, openApi: spec });
     expect(spy).toHaveBeenCalledWith(spec);
 
-    const spec2: Parameters<typeof parse>[0] = {
+    const spec2: OpenApi = {
       info: {
         title: 'dummy',
         version: '1.0',
@@ -68,10 +65,10 @@ describe('parse', () => {
       openapi: '3.0',
       paths: {},
     };
-    parse(spec2);
+    parse({ config, openApi: spec2 });
     expect(spy).toHaveBeenCalledWith(spec2);
 
-    const spec3: Parameters<typeof parse>[0] = {
+    const spec3: OpenApi = {
       info: {
         title: 'dummy',
         version: '1.0',
@@ -79,13 +76,13 @@ describe('parse', () => {
       openapi: '3.1.0',
       paths: {},
     };
-    parse(spec3);
+    parse({ config, openApi: spec3 });
     expect(spy).toHaveBeenCalledWith(spec3);
   });
 
   it('throws on unknown version', () => {
     // @ts-ignore
-    expect(() => parse({ foo: 'bar' })).toThrow(
+    expect(() => parse({ config, openApi: { foo: 'bar' } })).toThrow(
       `Unsupported Open API specification: ${JSON.stringify({ foo: 'bar' }, null, 2)}`,
     );
   });
