+++ packages/openapi-ts/src/openApi/common/interfaces/config.ts
@@ -1,0 +1,13 @@
+import type { Operation, OperationParameter } from './client';
+
+export interface Config {
+  debug?: boolean;
+  filterFn?: {
+    operation?: (operation: Operation) => boolean;
+    operationParameter?: (parameter: OperationParameter) => boolean;
+  };
+  nameFn: {
+    operation: (operation: Omit<Operation, 'name'>) => string;
+    operationParameter: (parameter: Omit<OperationParameter, 'name'>) => string;
+  };
+}
