+++ packages/openapi-ts/src/openApi/common/parser/__tests__/type.spec.ts
@@ -1,8 +1,7 @@
-import { describe, expect, it, type MockedFunction, vi } from 'vitest';
+import { describe, expect, it, vi } from 'vitest';
 
 import type { Config } from '../../../../types/config';
-import { isStandaloneClient } from '../../../../utils/config';
-import { getMappedType, getType, transformTypeKeyName } from '../type';
+import { getMappedType, getType } from '../type';
 
 vi.mock('../../../../utils/config', () => {
   const config: Partial<Config> = {
@@ -141,53 +140,5 @@ describe('getType', () => {
     expect(type.template).toEqual(null);
     expect(type.imports).toEqual([]);
     expect(type.isNullable).toEqual(true);
-  });
-});
-
-describe('transformTypeKeyName', () => {
-  describe('legacy client', () => {
-    it.each([
-      { expected: '', input: '' },
-      { expected: 'foobar', input: 'foobar' },
-      { expected: 'fooBar', input: 'fooBar' },
-      { expected: 'fooBar', input: 'foo_bar' },
-      { expected: 'fooBar', input: 'foo-bar' },
-      { expected: 'fooBar', input: 'foo.bar' },
-      { expected: 'fooBar', input: '@foo.bar' },
-      { expected: 'fooBar', input: '$foo.bar' },
-      { expected: 'fooBar', input: '123.foo.bar' },
-      { expected: 'fooBar', input: 'Foo-Bar' },
-      { expected: 'fooBar', input: 'FOO-BAR' },
-      { expected: 'fooBar', input: 'foo[bar]' },
-      { expected: 'fooBarArray', input: 'foo.bar[]' },
-    ])('$input -> $expected', ({ input, expected }) => {
-      (isStandaloneClient as MockedFunction<any>).mockImplementationOnce(
-        () => false,
-      );
-      expect(transformTypeKeyName(input)).toBe(expected);
-    });
-  });
-
-  describe('standalone client', () => {
-    it.each([
-      { expected: '', input: '' },
-      { expected: 'foobar', input: 'foobar' },
-      { expected: 'fooBar', input: 'fooBar' },
-      { expected: 'fooBar', input: 'foo_bar' },
-      { expected: 'fooBar', input: 'foo-bar' },
-      { expected: 'fooBar', input: 'foo.bar' },
-      { expected: 'fooBar', input: '@foo.bar' },
-      { expected: 'fooBar', input: '$foo.bar' },
-      { expected: 'fooBar', input: '123.foo.bar' },
-      { expected: 'fooBar', input: 'Foo-Bar' },
-      { expected: 'fooBar', input: 'FOO-BAR' },
-      { expected: 'fooBar', input: 'foo[bar]' },
-      { expected: 'fooBarArray', input: 'foo.bar[]' },
-    ])('$input -> $input', ({ input }) => {
-      (isStandaloneClient as MockedFunction<any>).mockImplementationOnce(
-        () => true,
-      );
-      expect(transformTypeKeyName(input)).toBe(input);
-    });
   });
 });
