+++ packages/openapi-ts/src/openApi/common/parser/type.ts
@@ -1,14 +1,8 @@
-import { camelCase } from '../../../utils/camelCase';
-import { getConfig, isStandaloneClient } from '../../../utils/config';
 import { refParametersPartial } from '../../../utils/const';
-import { reservedWordsRegExp } from '../../../utils/reservedWords';
 import { transformTypeName } from '../../../utils/transform';
 import { isDefinitionTypeNullable } from '../../v3/parser/inferType';
 import type { Type } from '../interfaces/Type';
-import {
-  ensureValidTypeScriptJavaScriptIdentifier,
-  sanitizeOperationParameterName,
-} from './sanitize';
+import { ensureValidTypeScriptJavaScriptIdentifier } from './sanitize';
 import { stripNamespace } from './stripNamespace';
 
 /**
@@ -171,22 +165,4 @@ export const getType = ({
   }
 
   return result;
-};
-
-/**
- * Replaces any invalid characters from a parameter name.
- * For example: 'filter.someProperty' becomes 'filterSomeProperty'.
- */
-export const transformTypeKeyName = (value: string): string => {
-  const config = getConfig();
-
-  // do not transform anything for standalone clients
-  if (isStandaloneClient(config)) {
-    return value;
-  }
-
-  const name = camelCase({
-    input: sanitizeOperationParameterName(value),
-  }).replace(reservedWordsRegExp, '_$1');
-  return name;
 };
