+++ packages/openapi-ts/src/openApi/index.ts
@@ -1,23 +1,34 @@
 import type { Client } from '../types/client';
+import type { Config } from './common/interfaces/config';
 import { OpenApi } from './common/interfaces/OpenApi';
+import { setConfig } from './config';
 import { parse as parseV2 } from './v2/index';
 import { parse as parseV3 } from './v3/index';
 
-export {
+export type {
   Enum,
   Model,
   Operation,
   OperationParameter,
   Service,
 } from './common/interfaces/client';
-export { OpenApi } from './common/interfaces/OpenApi';
+export type { Config } from './common/interfaces/config';
+export type { OpenApi } from './common/interfaces/OpenApi';
 
 /**
  * Parse the OpenAPI specification to a Client model that contains
  * all the models, services and schema's we should output.
  * @param openApi The OpenAPI spec that we have loaded from disk.
  */
-export function parse(openApi: OpenApi): Client {
+export function parse({
+  openApi,
+  config,
+}: {
+  config: Config;
+  openApi: OpenApi;
+}): Client {
+  setConfig(config);
+
   if ('openapi' in openApi) {
     return parseV3(openApi);
   }
