+++ packages/openapi-ts/src/openApi/v2/index.ts
@@ -14,11 +14,10 @@ export const parse = (openApi: OpenApi): Client => {
   const version = getServiceVersion(openApi.info.version);
   const server = getServer(openApi);
   const { models, types } = getModels(openApi);
-  const { operationIds, services } = getServices({ openApi, types });
+  const services = getServices({ openApi, types });
 
   return {
     models,
-    operationIds,
     server,
     services,
     types,
