+++ packages/openapi-ts/src/openApi/v2/parser/operation.ts
@@ -3,12 +3,10 @@ import type {
   Operation,
   OperationParameters,
 } from '../../common/interfaces/client';
-import {
-  getOperationName,
-  getOperationResponseHeader,
-} from '../../common/parser/operation';
+import { getOperationResponseHeader } from '../../common/parser/operation';
 import { getServiceName } from '../../common/parser/service';
 import { toSortedByRequired } from '../../common/parser/sort';
+import { getConfig } from '../../config';
 import type { OpenApi } from '../interfaces/OpenApi';
 import type { OpenApiOperation } from '../interfaces/OpenApiOperation';
 import { getOperationParameters } from './getOperationParameters';
@@ -31,17 +29,16 @@ export const getOperation = ({
   types: Client['types'];
   url: string;
 }): Operation => {
+  const config = getConfig();
   const service = getServiceName(tag);
-  const name = getOperationName(url, method, op.operationId);
 
-  const operation: Operation = {
+  const operationWithoutName: Omit<Operation, 'name'> = {
     $refs: [],
     deprecated: op.deprecated === true,
     description: op.description || null,
     id: op.operationId || null,
     imports: [],
     method: method.toUpperCase() as Operation['method'],
-    name,
     parameters: [...pathParams.parameters],
     parametersBody: pathParams.parametersBody,
     parametersCookie: [...pathParams.parametersCookie],
@@ -54,6 +51,10 @@ export const getOperation = ({
     responses: [],
     service,
     summary: op.summary || null,
+  };
+  const operation = {
+    ...operationWithoutName,
+    name: config.nameFn.operation(operationWithoutName),
   };
 
   if (op.parameters) {
