+++ packages/openapi-ts/src/openApi/v3/parser/getModels.ts
@@ -1,6 +1,6 @@
 import type { Client } from '../../../types/client';
-import { getConfig } from '../../../utils/config';
 import { getParametersMeta, getSchemasMeta } from '../../../utils/meta';
+import { getConfig } from '../../config';
 import type { OpenApi } from '../interfaces/OpenApi';
 import { getModel } from './getModel';
 import { getParameterSchema } from './parameter';
