+++ packages/openapi-ts/src/openApi/v3/parser/getOperationParameters.ts
@@ -1,7 +1,7 @@
 import type { Client } from '../../../types/client';
-import { getConfig, isStandaloneClient } from '../../../utils/config';
 import type { OperationParameters } from '../../common/interfaces/client';
 import { getRef } from '../../common/parser/getRef';
+import { getConfig } from '../../config';
 import type { OpenApi } from '../interfaces/OpenApi';
 import type { OpenApiParameter } from '../interfaces/OpenApiParameter';
 import { getOperationParameter } from './getOperationParameter';
@@ -9,19 +9,15 @@ import { getOperationParameter } from './getOperationParameter';
 const allowedIn = ['cookie', 'formData', 'header', 'path', 'query'] as const;
 
 export const getOperationParameters = ({
-  debug,
   openApi,
   parameters,
   types,
 }: {
-  debug?: boolean;
   openApi: OpenApi;
   parameters: OpenApiParameter[];
   types: Client['types'];
 }): OperationParameters => {
   const config = getConfig();
-
-  const isStandalone = isStandaloneClient(config);
 
   const operationParameters: OperationParameters = {
     $refs: [],
@@ -41,18 +37,15 @@ export const getOperationParameters = ({
       parameterOrReference,
     );
     const parameter = getOperationParameter({
-      debug,
       openApi,
       parameter: parameterDef,
       types,
     });
 
-    // legacy clients ignore the "api-version" param since we do not want to
-    // add it as the first/default parameter for each of the service calls
-    if (
-      !allowedIn.includes(parameterDef.in) ||
-      (!isStandalone && parameter.prop === 'api-version')
-    ) {
+    const skip =
+      config.filterFn?.operationParameter &&
+      !config.filterFn?.operationParameter(parameter);
+    if (!allowedIn.includes(parameterDef.in) || skip) {
       return;
     }
