+++ packages/openapi-ts/src/utils/type.ts
@@ -1,12 +1,14 @@
 import { compiler, type Property, type TypeNode } from '../compiler';
 import type { Model } from '../openApi';
-import { transformTypeKeyName } from '../openApi/common/parser/type';
+import { sanitizeOperationParameterName } from '../openApi/common/parser/sanitize';
 import type { Client } from '../types/client';
+import { camelCase } from './camelCase';
 import { getConfig, isStandaloneClient } from './config';
 import { refSchemasPartial } from './const';
 import { enumValue } from './enum';
 import { escapeComment, escapeName, unescapeName } from './escape';
 import { getSchemasMeta } from './meta';
+import { reservedWordsRegExp } from './reservedWords';
 import { unique } from './unique';
 
 const base = (model: Model) => {
@@ -317,4 +319,22 @@ export const unsetUniqueTypeName = ({
     name,
   };
   return result;
+};
+
+/**
+ * Replaces any invalid characters from a parameter name.
+ * For example: 'filter.someProperty' becomes 'filterSomeProperty'.
+ */
+export const transformTypeKeyName = (value: string): string => {
+  const config = getConfig();
+
+  // do not transform anything for standalone clients
+  if (isStandaloneClient(config)) {
+    return value;
+  }
+
+  const name = camelCase({
+    input: sanitizeOperationParameterName(value),
+  }).replace(reservedWordsRegExp, '_$1');
+  return name;
 };
